Refactor the backend impl package
