+++ backend/impl/src/test/java/com/virtuslab/gitmachete/backend/unit/BaseGitMacheteRepositoryUnitTestSuite.java
@@ -7,23 +7,14 @@ import io.vavr.collection.List;
 import lombok.SneakyThrows;
 import lombok.val;
 import org.powermock.api.mockito.PowerMockito;
-import org.powermock.reflect.Whitebox;
 
 import com.virtuslab.gitcore.api.IGitCoreBranchSnapshot;
 import com.virtuslab.gitcore.api.IGitCoreHeadSnapshot;
 import com.virtuslab.gitcore.api.IGitCoreRepository;
-import com.virtuslab.gitmachete.backend.impl.GitMacheteRepository;
-import com.virtuslab.gitmachete.backend.impl.hooks.PreRebaseHookExecutor;
+import com.virtuslab.gitmachete.backend.impl.aux.CreateGitMacheteRepositoryAux;
 import com.virtuslab.gitmachete.backend.impl.hooks.StatusBranchHookExecutor;
 
 public class BaseGitMacheteRepositoryUnitTestSuite {
-
-  private static final Class<?> AUX_CLASS = getAuxClass();
-
-  @SneakyThrows
-  private static Class<?> getAuxClass() {
-    return Whitebox.getInnerClassType(GitMacheteRepository.class, "CreateGitMacheteRepositoryAux");
-  }
 
   protected final IGitCoreRepository gitCoreRepository = PowerMockito.mock(IGitCoreRepository.class);
 
@@ -41,8 +32,6 @@ public class BaseGitMacheteRepositoryUnitTestSuite {
     // cannot be mocked as it is final
     val statusBranchHookExecutor = StatusBranchHookExecutor.of(gitCoreRepository);
 
-    return Whitebox
-        .getConstructor(AUX_CLASS, IGitCoreRepository.class, StatusBranchHookExecutor.class, PreRebaseHookExecutor.class)
-        .newInstance(gitCoreRepository, statusBranchHookExecutor, /* preRebaseHookExecutor */ null);
+    return new CreateGitMacheteRepositoryAux(gitCoreRepository, statusBranchHookExecutor, /* preRebaseHookExecutor */ null);
   }
 }
